+++ routes/resources.js
@@ -1,57 +1,13 @@
 'use strict';
 var P = require('bluebird');
-var _ = require('lodash');
-var logger = require('../services/logger');
-var ResourceSerializer = require('../serializers/resource');
-var ResourceDeserializer = require('../deserializers/resource');
 var auth = require('../services/auth');
 var path = require('../services/path');
-var Schemas = require('../generators/schemas');
+var ResourceSerializer = require('../serializers/resource');
+var ResourceDeserializer = require('../deserializers/resource');
+var SmartFieldsValuesInjector = require('../services/smart-fields-values-injector');
 
 module.exports = function (app, model, Implementation, integrator, opts) {
   var modelName = Implementation.getModelName(model);
-  var schema = Schemas.schemas[modelName];
-
-  function setSmartFieldValue(record, field, modelName) {
-    if (field.value) {
-      logger.warn('DEPRECATION WARNING: Smart Fields "value" method is ' +
-        'deprecated. Please use "get" method in your collection ' +
-        modelName + ' instead.');
-    }
-
-    var value = field.get ? field.get(record) : field.value(record);
-    if (value && _.isFunction(value.then)) {
-      return value.then(function (result) { record[field.field] = result; });
-    } else {
-      record[field.field] = value;
-    }
-  }
-
-  function injectSmartFields(record) {
-    return P.each(schema.fields, function (field) {
-      if (!record[field.field]) {
-        if (field.get || field.value) {
-          return setSmartFieldValue(record, field, modelName);
-        } else if (_.isArray(field.type)) {
-          record[field.field] = [];
-        }
-      } else if (field.reference && !_.isArray(field.type)) {
-        // NOTICE: Set Smart Fields values to "belongsTo" associated records.
-        var modelNameAssociation = field.reference.split('.')[0];
-        var schemaAssociation = Schemas.schemas[modelNameAssociation];
-
-        if (schemaAssociation) {
-          return P.each(schemaAssociation.fields, function (fieldAssociation) {
-            if (!record[field.field][fieldAssociation.field] &&
-              (fieldAssociation.get || fieldAssociation.value)) {
-              return setSmartFieldValue(record[field.field],
-                fieldAssociation, modelNameAssociation);
-            }
-          });
-        }
-      }
-    }).thenReturn(record);
-  }
 
   this.list = function (req, res, next) {
     return new Implementation.ResourcesGetter(model, opts, req.query)
@@ -60,9 +16,10 @@ module.exports = function (app, model, Implementation, integrator, opts) {
         var count = results[0];
         var records = results[1];
 
-        // Inject smart fields.
         return P
-          .map(records, injectSmartFields)
+          .map(records, function (record) {
+            return new SmartFieldsValuesInjector(record, modelName).perform();
+          })
           .then(function (records) {
             return new ResourceSerializer(Implementation, model, records,
               integrator, opts, { count: count }).perform();
@@ -77,7 +34,9 @@ module.exports = function (app, model, Implementation, integrator, opts) {
   this.get = function (req, res, next) {
     return new Implementation.ResourceGetter(model, req.params)
       .perform()
-      .then(injectSmartFields)
+      .then(function(records) {
+        return new SmartFieldsValuesInjector(records, modelName).perform();
+      })
       .then(function (record) {
         return new ResourceSerializer(Implementation, model, record,
           integrator, opts).perform();